requirements.txt
# QQ聊天机器人长期记忆系统 - 依赖包列表

# ========== Web框架 ==========
flask>=2.0.0
flask-cors>=3.0.0

# ========== 配置文件解析 ==========
pyyaml>=6.0

# ========== 深度学习框架 ==========
torch>=2.0.0
transformers>=4.30.0

# ========== 模型加载 ==========
modelscope>=1.0.0

# ========== 参数高效微调 ==========
peft>=0.5.0

# ========== 分布式训练和混合精度 ==========
accelerate>=0.20.0

# ========== HTTP请求 ==========
requests>=2.28.0

# ========== 定时任务调度 ==========
apscheduler>=3.10.0

# ========== 进度条 ==========
tqdm>=4.64.0

# ========== 图像处理 ==========
Pillow>=9.0.0

# ========== 文件处理（可选）==========
# PDF文本提取
pdfminer.six>=20221105
# PDF转图片（需要系统安装poppler-utils）
pdf2image>=1.16.0
# DOCX文件处理
python-docx>=0.8.11
# OCR（可选，需要系统安装tesseract）
pytesseract>=0.3.10
# 网页内容提取（可选）
beautifulsoup4>=4.11.0

# ========== 客户端依赖（如果需要在服务器上运行客户端）==========
# ncatbot>=0.1.0  # 通常只在客户端需要，服务器端不需要

# ========== 开发和测试工具（可选）==========
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0

